+++ src/ts/Ads/Managers/PerPlacementLoadManagerV5.ts
@@ -18,6 +18,7 @@ export class PerPlacementLoadManagerV5 extends PerPlacementLoadManager {
     protected _adRequestManager: AdRequestManager;
 
     private _shouldRefresh: boolean = true;
+    private _lastShownCampaignId: string | undefined;
 
     constructor(ads: IAdsApi, adsConfig: AdsConfiguration, coreConfig: CoreConfiguration, adRequestManager: AdRequestManager, clientInfo: ClientInfo, focusManager: FocusManager, useGroupIds: boolean) {
         super(ads, adsConfig, coreConfig, useGroupIds ? new AdUnitAwareAdRequestManager(adRequestManager) : adRequestManager, clientInfo, focusManager);
@@ -30,6 +31,11 @@ export class PerPlacementLoadManagerV5 extends PerPlacementLoadManager {
 
     public setCurrentAdUnit(adUnit: AbstractAdUnit, placement: Placement): void {
         this._shouldRefresh = true;
+
+        const campaign = placement.getCurrentCampaign();
+        if (campaign) {
+            this._lastShownCampaignId = campaign.getUniqueId();
+        }
 
         Observables.once(adUnit.onStartProcessed, () => {
             // Aids in supplying comet a suitable amount of time to process start event
@@ -157,7 +163,8 @@ export class PerPlacementLoadManagerV5 extends PerPlacementLoadManager {
             placement.setInvalidationPending(false);
 
             let shouldInvalidate = true;
-            // Invalidate only Direct Demand campaigns, we would like to show old programmatic campaign.
+
+                // Invalidate only Direct Demand campaigns, we would like to show old programmatic campaign.
             const campaign = placement.getCurrentCampaign();
             if (campaign) {
                 const contentType = campaign.getContentType();
@@ -169,6 +176,11 @@ export class PerPlacementLoadManagerV5 extends PerPlacementLoadManager {
                         break;
                     default:
                         shouldInvalidate = false;
+                }
+
+                if (!shouldInvalidate && this._lastShownCampaignId && campaign.getUniqueId() === this._lastShownCampaignId) {
+                    this._adRequestManager.reportMetricEvent(LoadV5.RefreshManagerForcedToInvalidate);
+                    shouldInvalidate = true;
                 }
             }